Add files via upload
Try finally begin. — Max 
With the following report I show how to host and execute a deep learning project on a cloud. The cloud is hosted by google colab and enables working and testing in teams. 
Lazarus is also being built in colab and the deep learning network is compiled and trained too in a Jupiter notebook.